+++ crates/core/tests/java/annotation_new_line_in_record_component/component_wrap_self.java
@@ -1,0 +1,8 @@
+record R(
+        @Deprecated
+        int alphaComponent,
+        @A
+        @B
+        String betaComponent,
+        int gammaComponent
+) {}